+++ CHANGELOG.md
@@ -36,6 +36,8 @@
   * Allow weapon status check for dice roll also for shipweapon
   * Blind GM rolls are now properly hidden from players ([#1712](https://github.com/StarWarsFoundryVTT/StarWarsFFG/issues/1712))
   * Properly use Genesys story points when the theme is set to Genesys ([#1711](https://github.com/StarWarsFoundryVTT/StarWarsFFG/issues/1711))
+  * Stop using deprecated `select` Handlebars helper ([#1632](https://github.com/StarWarsFoundryVTT/StarWarsFFG/issues/1632))
+    * Note that the same _code_ is in use, it is just no longer using the built-in (and deprecated) version
 
 `1.903`
 * Features: